Adding handling of unexpected messages in NT-mode, if no process exists.
Sometimes the call is released before a telephone responds.
This can happen, on long l1 activation phase, or if caller hangs up before a
telehone answers, or if one telephone connects before other phones responds.
In this case, the phone gets released with cause 81 (Invalid Callref).

This conforms to ETSI!