Some code style helper
I would prefer, if the source code for the mISDNuser project
follow the some common style.

This is the my prefered option but I also accept others - but maybe I will
reformat it.

It is nearly the same style the Linux kernel is using, with the exception
that I allow 132 characters per line (kernel has a strict 80 character limit).

To get your file reformatted with this style, simple run this file as script:
./CodeStyle <path to source file>

The script part was copied from the linux kernel sources.

Signed-off-by: Karsten Keil <kkeil@linux-pingi.de>