Only the highest matched bit position gives the CIP value
CAPI 2.0 spec defines the CIP value indicated by CONNECT_IND
als highest matched bit position, not the higest CIP value
calculated from bearer + HLC info.
This bug caused that if a application did not request listening
for some of the higher bit values a not requested CIP was sent, so the
call was ignored.

Signed-off-by: Karsten Keil <keil@b1-systems.de>