mISDNcapid: Fix INFO_IND IE posting function early return condition
lPLCIInfoIndIE() in lplci.c returned early when mc->l3m was set and
continued execution when it was NULL, however from code further in this
function it is clear that opposite behavior was meant.

This caused CAPI applications to not receive INFO_IND with information
elements, as this function was no-op when called correctly.

Signed-off-by: Maciej S. Szmigiero <mail@maciej.szmigiero.name>