mISDNcapid: Parse B channel id first from message then its included IEs
If early B3 is enabled we are currently opening channel B on appropriate
"Progress" IE.

If it turns out that this IE arrives in the same message that also sets
which B channel should be used the current code will fail to open
the B channel for early B3 as the current order is to first handle the IEs
then parse channel id.

Change this to first parse the B channel id then handle the IEs so in such
situation the B channel id will already be set at "Progress" IE handling
time.

Signed-off-by: Maciej S. Szmigiero <mail@maciej.szmigiero.name>