mISDNcapid: remove irritating timeout warning
The loop within BCthread() continuously emits a warning like the following one

  BCthread(24062):Bchannel1 timeout (release not pending) thread=24062

every 500 ms as long as no data is received over the B-channel. This is very
irritating and noisy and fills up the log file without a good reason. This
commit lets the warning be shown only if the timeout occurs while a release
operation is pending (and this is worth a warning because such a timeout does
not occur under normal circumstances).

Signed-off-by: Christoph Schulz <develop@kristov.de>